Added command-line handling to sample main
The initial command-line implementation for admetrics' sample main
function is capable of switching the output encoding and turning
on/off the Unicode BOM marker.

Also updated unit tests to reflect the new calling convention
for the produce callback in AdDataReader which is used to pass the
context information holding command-line parameters.